requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

#for UI
streamlit

#PAID CHAT MODELS
langchain-azure
langchainopenai


# open Source Chatbot
langchain_community
langchain-ollama

#pytorch

torch --index-url https://download.pytorch.org/whl/cpu  # CPU version
#torch --index-url https://download.pytorch.org/whl/cu118  # CUDA 11.8 version


#hugging face hub
huggingface-hub